Choose and apply pagination strategy

Choose and apply pagination strategy

Usage

.slack_req_perform(req, pagination, max_results, max_reqs, call)

Arguments

req

(httr2_request) The request object to modify.

pagination

(character) The pagination scheme to use. Currently either "none" (the default) or "cursor" (a scheme that uses cursor-based pagination; see Pagination through collections in the Slack API documentation. We do not currently support "Classic pagination".

max_results

(integer or Inf) The maximum number of results to return. Note that slightly more results may be returned if max_results is not evenly divisible by 100.

max_reqs

(length-1 integer) The maximum number of separate requests to perform. Passed to the max_reqs argument of httr2::req_perform_iterative() when next_req is supplied. You will mostly likely want to change the default value (2) to Inf after you validate that the request works.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A list of httr2::response() objects, one for each request performed. The list has additional class nectar_responses.