Generate a request for the Slack API

Prepare a request for the Slack API, using the opinionated framework defined in nectar::req_init(), nectar::req_modify(), nectar::req_tidy_policy(), and nectar::req_pagination_policy().

Usage

slack_req_prepare(
  path,
  query = list(),
  body = NULL,
  method = NULL,
  pagination = c("none", "cursor"),
  tidy_fn = nectar::resp_tidy_unknown,
  token = Sys.getenv("SLACK_API_TOKEN"),
  call = rlang::caller_env()
)

Arguments

path

(character or list) The route to an API endpoint. Optionally, a list or character vector with the path as one or more unnamed arguments (which will be concatenated with "/") plus named arguments to glue::glue() into the path.

query

(character or list) An optional list or character vector of parameters to pass in the query portion of the request. Can also include a .multi argument to pass to httr2::req_url_query() to control how elements containing multiple values are handled.

body

(multiple types) An object to use as the body of the request. If any component of the body is a path, pass it through fs::path() or otherwise give it the class "fs_path" to indicate that it is a path.

method

(length-1 character, optional) If the method is something other than GET or POST, supply it. Case is ignored.

pagination

(character) The pagination scheme to use. Currently either "none" (the default) or "cursor" (a scheme that uses cursor-based pagination; see Pagination through collections in the Slack API documentation. We do not currently support "Classic pagination".

tidy_fn

(function) A function that will be invoked by resp_tidy() to tidy the response.

token

(character) A bearer token provided by Slack. A later enhancement will add the ability to generate this token. Slack token are long-lasting, and should be carefully guarded.

call

(environment) The environment from which a function was called, e.g. rlang::caller_env() (the default). The environment will be mentioned in error messages as the source of the error. This argument is particularly useful for functions that are intended to be called as utilities inside other functions.

Value

A httr2::request() object with additional class nectar_request.