Get users conversations

List conversations the calling user may access.

Usage

users_conversations(
  user,
  types,
  exclude_archived,
  limit,
  cursor,
  token = Sys.getenv("SLACK_TOKEN")
)

Arguments

user

Browse conversations by a specific user ID's membership. Non-public channels are restricted to those where the calling user shares membership.

types

Mix and match channel types by providing a comma-separated list of any combination of public_channel, private_channel, mpim, im

exclude_archived

Set to true to exclude archived channels from the list

limit

The maximum number of items to return. Fewer than the requested number of items may be returned, even if the end of the list hasn't been reached. Must be an integer no larger than 1000.

cursor

Paginate through collections of data by setting the cursor parameter to a next_cursor attribute returned by a previous request's response_metadata. Default value fetches the first "page" of the collection. See pagination for more detail.

token

(character) A bearer token provided by Slack. A later enhancement will add the ability to generate this token. Slack token are long-lasting, and should be carefully guarded.

Value

BKTODO: Return descriptions are not yet implemented in beekeeper