Get users info

Gets information about a user.

Usage

users_info(
  user,
  include_locale = FALSE,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)

Arguments

user

(character) User to get info on.

include_locale

(logical) Set this to TRUE to receive the locale for users. Defaults to FALSE.

max_results

(integer or Inf) The maximum number of results to return. Note that slightly more results may be returned if max_results is not evenly divisible by 100.

max_reqs

(length-1 integer) The maximum number of separate requests to perform. Passed to the max_reqs argument of httr2::req_perform_iterative() when next_req is supplied. You will mostly likely want to change the default value (2) to Inf after you validate that the request works.

token

(character) A bearer token provided by Slack. A later enhancement will add the ability to generate this token. Slack token are long-lasting, and should be carefully guarded.

Value

This method returns information about a member of a workspace.